misc/HtmlTestSuite/results/test-0540.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Grid Layout Test: parsing grid-template-rows with valid values</title>
<link href="https://drafts.csswg.org/css-grid-1/#propdef-grid-template-rows" rel="help"/>
<meta content="grid-template-rows supports the full grammar 'none | &lt;track-list&gt; | &lt;auto-track-list&gt;'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_valid_value("grid-template-rows", "none");

// track-size &lt;fixed-breadth&gt; = &lt;length-percentage&gt; | &lt;flex&gt; | min-content | max-content | auto
test_valid_value("grid-template-rows", '10px');
test_valid_value("grid-template-rows", '20%');
test_valid_value("grid-template-rows", 'calc(-0.5em + 10px)');
test_valid_value("grid-template-rows", 'calc(0.5em + 10px)');
test_valid_value("grid-template-rows", 'calc(30% + 40vw)');
test_valid_value("grid-template-rows", '5fr');
test_valid_value("grid-template-rows", 'min-content');
test_valid_value("grid-template-rows", 'max-content');
test_valid_value("grid-template-rows", 'auto');

// track-size minmax( &lt;inflexible-breadth&gt; , &lt;track-breadth&gt; )
test_valid_value("grid-template-rows", 'minmax(10px, auto)');
test_valid_value("grid-template-rows", 'minmax(20%, max-content)');
test_valid_value("grid-template-rows", 'minmax(calc(-0.5em + 10px), min-content)');
test_valid_value("grid-template-rows", 'minmax(calc(0.5em + 10px), 5fr)');
test_valid_value("grid-template-rows", 'minmax(calc(30% + 40vw), 10px)');
test_valid_value("grid-template-rows", 'minmax(min-content, 20%)');
test_valid_value("grid-template-rows", 'minmax(max-content, calc(-0.5em + 10px))');
test_valid_value("grid-template-rows", 'minmax(auto, calc(0.5em + 10px))');

// track-size fit-content( &lt;length-percentage&gt; )
test_valid_value("grid-template-rows", 'fit-content(10px)');
test_valid_value("grid-template-rows", 'fit-content(20%)');
test_valid_value("grid-template-rows", 'fit-content(calc(-0.5em + 10px))');
test_valid_value("grid-template-rows", 'fit-content(calc(0.5em + 10px))');
test_valid_value("grid-template-rows", 'fit-content(calc(30% + 40vw))');

// &lt;track-repeat&gt; = repeat( [ &lt;positive-integer&gt; ] , [ &lt;line-names&gt;? &lt;track-size&gt; ]+ &lt;line-names&gt;? )

test_valid_value("grid-template-rows", 'repeat(1, [] 10px [])', 'repeat(1, 10px)');

// 'repeat(1, [one two] 20%)' in Blink, Firefox
// '[one two] 20%' in Safari, Edge 18
test_valid_value("grid-template-rows", 'repeat(1, [one two] 20%)');

// 'repeat(2, minmax(10px, auto))' in Blink, Firefox, Edge 18
// 'minmax(10px, auto) minmax(10px, auto)' in Safari
test_valid_value("grid-template-rows", 'repeat(2, minmax(10px, auto))');

test_valid_value("grid-template-rows", 'repeat(2, fit-content(20%) [three four] 30px 40px [five six])');

// &lt;track-list&gt; = [ &lt;line-names&gt;? [ &lt;track-size&gt; | &lt;track-repeat&gt; ] ]+ &lt;line-names&gt;?
test_valid_value("grid-template-rows", 'min-content repeat(5, minmax(10px, auto))');
test_valid_value("grid-template-rows", '[] 150px [] 1fr []', '150px 1fr');

// &lt;auto-repeat&gt; = repeat( [ auto-fill | auto-fit ] , [ &lt;line-names&gt;? &lt;fixed-size&gt; ]+ &lt;line-names&gt;? )
test_valid_value("grid-template-rows", 'repeat(auto-fill, 10px)');
test_valid_value("grid-template-rows", 'repeat(auto-fit, [one] 20%)');
test_valid_value("grid-template-rows", 'repeat(auto-fill, minmax(30px, 5fr) [two])');
test_valid_value("grid-template-rows", 'repeat(auto-fit, [three] minmax(max-content, 6em) [four])');

// &lt;auto-track-list&gt; =
// [ &lt;line-names&gt;? [ &lt;fixed-size&gt; | &lt;fixed-repeat&gt; ] ]* &lt;line-names&gt;?
// &lt;auto-repeat&gt;
// [ &lt;line-names&gt;? [ &lt;fixed-size&gt; | &lt;fixed-repeat&gt; ] ]* &lt;line-names&gt;?

test_valid_value("grid-template-rows", '[one] repeat(2, minmax(10px, auto)) [two] 30px [three] repeat(auto-fill, 10px) 40px [four five] repeat(2, minmax(10px, auto)) [six]');
</script>
</body>
</html>
